+++ particle.h
@@ -32,7 +32,7 @@ void DrawParticles();
 void InitParticles();
 void AddParticle(float x, float y, short lifetime, float speed, float angle, Color colour); 
 void AddMouseRandomParticles(MouseState mouseState, int numParticles);
-void RandParticleAroundEdgeOfCircle(float cx, float cy, float r, short lifetime, float maxspeed, Color col);
+void RandParticleAroundEdgeOfCircle(float cx, float cy, float r, float lifetime, float maxspeed, Color col);
 void AddParticleWithRandomProperties(float x, float y, Color color, float lifetimeMin, float lifetimeMax, float speedMin, float speedMax, float minAngle, float maxAngle);
 bool ParticleIsActive(int index);
 float GetParticleAlpha(int index);
